+++ changelog.md
@@ -6,6 +6,9 @@ This change log records updates after 2025-3-24.
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
 ## [v0.5.1] - 2025-4-17
+### Added
+- Add `SimpleRNN` layer initial implementation
+
 ### Changed
 - Use rand v0.9.1